Correct return type for TPTScriptInterface::parseInt
so that large integers don't get changed into slightly different values by rounding caused by conversion to a float and back again.

+++ src/lua/TPTScriptInterface.cpp
@@ -124,7 +124,7 @@ ValueType TPTScriptInterface::testType(std::string word)
 			return TypeString;
 }
 
-float TPTScriptInterface::parseNumber(char * stringData)
+int TPTScriptInterface::parseNumber(char * stringData)
 {
 	char cc;
 	int base = 10;
